Updated MentorWizard to mirror ParentWizard
